Generic S3 method for configuring the default prior for a bmmmodel

Called by fit_model() to automatically construct the priors for a given model, data and formula, and combine it with the prior given by the user. The first method executed is configure_prior.bmmmodel, which will build the prior based on information from the model object such as fixed_parameters, default_priors, etc. Thus it is important to define these values in the model object. The function will also recognize if the user has specified that some parameters should be fixed to a constant and put the appropriate constant priors. Any additional priors that a developer wants to specify, which are not based on information in the model object, can be defined in the configure_prior.* method for the model. See configure_prior.IMMfull for an example.

Usage

configure_prior(model, data, formula, user_prior, ...)

Arguments

model

A bmmmodel object

data

A data.frame containing the data used in the model

formula

A brmsformula object returned from configure_model()

user_prior

A brmsprior object given by the user as an argument to fit_model()

...

Additional arguments passed to the method