Fit Measurement Models using BRMS

Fit a Bayesian multilevel measurement model. Currently implemented are the two-parameter mixture model by Zhang and Luck (2008), the three-parameter mixture model by Bays et al (2009), and three different versions of the Interference Measurement Model (Oberauer et al., 2017). This is a wrapper function for brms::brm, which is used to estimate the model.

Usage

fit_model(
  formula,
  data,
  model,
  parallel = FALSE,
  chains = 4,
  prior = NULL,
  ...
)

Arguments

formula

An object of class brmsformula. A symbolic description of the model to be fitted.

data

An object of class data.frame, containing data of all variables used in the model. The names of the variables must match the variable names passed to the bmmmodel object for required argurments.

model

A description of the model to be fitted. This is a call to a bmmmodel such as mixture3p() function. Every model function has a number of required arguments which need to be specified within the function call. Call supported_models() to see the list of supported models and their required arguments

parallel

Logical; If TRUE, the number of cores on your machine will be detected and brms will fit max(chains, cores) number of chains (specified by the chain argument) in parallel using the parallel package

chains

Numeric. Number of Markov chains (defaults to 4)

prior

One or more brmsprior objects created by brms::set_prior() or related functions and combined using the c method or the + operator. See also brms::get_prior() for more help. Not necessary for the default model fitting, but you can provide prior constraints to model parameters

...

Further arguments passed to brms::brm() or Stan. See the description of brms::brm() for more details

Value

An object of class brmsfit which contains the posterior draws along with many other useful information about the model. Use methods(class = "brmsfit") for an overview on available methods.

Details

The following models are supported:

IMMabc(non_targets, setsize)

IMMbsc(non_targets, setsize, spaPos)

IMMfull(non_targets, setsize, spaPos)

mixture2p()

mixture3p(non_targets, setsize)

sdmSimple()

Type ?modelname to get information about a specific model, e.g. ?IMMfull

Type help(package=bmm) for a full list of available help topics.

References

Frischkorn, G. T., & Popov, V. (2023). A tutorial for estimating mixture models for visual working memory tasks in brms: Introducing the Bayesian Measurement Modeling (bmm) package for R. https://doi.org/10.31234/osf.io/umt57

See also

Examples

if (FALSE) {
# generate artificial data from the Bays et al (2009) 3-parameter mixture model
dat <- gen_3p_data(N=2000, pmem=0.6, pnt=0.3, kappa=10, setsize=4, relative_resp=T)

# define formula
ff <- brms::bf(y ~ 1,
              kappa ~ 1,
              thetat ~ 1,
              thetant ~ 1)

# fit the model
fit <- fit_model(formula = ff,
                 data = dat,
                 model = mixture3p(non_targets = paste0('nt',1:3,'_loc'), setsize=4),
                 parallel=T,
                 iter=500,
                 backend='cmdstanr')
}