Generate artificial data from the Bays et al (2009) 3-parameter mixture model

Given a set of parameters, and desired setsize, it generates responses in a continuous reproduction task, with corresponding proportions coming from memory, non-target location errors, or guessing.

Usage

gen_3p_data(
  N = 2000,
  pmem = 0.6,
  pnt = 0.3,
  kappa = 10,
  setsize = 2,
  relative_resp = T
)

Arguments

N

Numeric. Number of samples

pmem

Numeric. Probability of response coming from memory

pnt

Numeric. Probability of response coming from a non-target item

kappa

Numeric. Precision of the von mises distribution

setsize

Numeric. Number of presented items

relative_resp

Logical. if TRUE, returns response error relative to the target and all non-target item locations are coded relative to the target. If FALSE, returns actual response and the location of all items is absolute

Value

A data.frame object. y is the response (if relative_resp==F) or response error (if relative_resp==T), t_loc is the value of the target (only if relative_resp==F), nt1_loc to nti_loc, are values of the non-targets, where i=setsize-1.

Examples

# example code
nsub = 30
df_3p_parms <- data.frame(
  theta_pmem = rnorm(nsub, mean = 2, sd = 1),
  theta_pnt = rnorm(nsub, mean = 1, sd = 0.5),
  kappa = pmax(0,rnorm(nsub, mean = 5, sd = 1)),
  pmem = numeric(nsub),
  pnt = numeric(nsub),
  pguess = numeric(nsub)
)

# transform continous mixture weights into probabilities using the softmax
df_3p_parms[,c("pmem","pnt","pguess")] <- apply(df_3p_parms[,c("theta_pmem","theta_pnt")],
                                                1,softmax)