Generate data for bmm models to be passed to Stan

A wrapper around brms::make_standata() for models specified with bmm. Given the model, the data and the formula for the model, this function will return the combined stan data generated by bmm and brms

Usage

get_standata(formula, data, model, prior = NULL, ...)

Arguments

formula

An object of class brmsformula. A symbolic description of the model to be fitted.

data

An object of class data.frame, containing data of all variables used in the model. The names of the variables must match the variable names passed to the bmmmodel object for required argurments.

model

A description of the model to be fitted. This is a call to a bmmmodel such as mixture3p() function. Every model function has a number of required arguments which need to be specified within the function call. Call supported_models() to see the list of supported models and their required arguments

prior

One or more brmsprior objects created by brms::set_prior() or related functions and combined using the c method or the + operator. See also get_model_prior() for more help. Not necessary for the default model fitting, but you can provide prior constraints to model parameters

...

Further arguments passed to brms::make_standata(). See the description of brms::make_standata() for more details

Value

A named list of objects containing the required data to fit a bmm model with Stan.

See also

Examples

if (FALSE) {
# generate artificial data from the Signal Discrimination Model
dat <- data.frame(y=rsdm(n=2000))

# define formula
ff <- bmf(c ~ 1,
          kappa ~ 1)

# fit the model
get_standata(formula = ff,
             data = dat,
             model = sdmSimple(resp_err = "y")
)
}