Control Saving of Parameter Draws

Thin wrapper around brms::save_pars(). When calling fit_model with additional information to save parameters you can use this function to pass information about saving parameter draws to brms without having to load brms. Alternatively, you can also load brms and call save_pars. For details see ?brms::save_pars.

Usage

save_pars(group = TRUE, latent = FALSE, all = FALSE, manual = NULL)

Arguments

group

A flag to indicate if group-level coefficients for each level of the grouping factors should be saved (default is TRUE). Set to FALSE to save memory. Alternatively, group may also be a character vector naming the grouping factors for which to save draws of coefficients.

latent

A flag to indicate if draws of latent variables obtained by using me and mi terms should be saved (default is FALSE). Saving these draws allows to better use methods such as posterior_predict with the latent variables but leads to very large R objects even for models of moderate size and complexity. Alternatively, latent may also be a character vector naming the latent variables for which to save draws.

all

A flag to indicate if draws of all variables defined in Stan's parameters block should be saved (default is FALSE). Saving these draws is required in order to apply the certain methods such as bridge_sampler and bayes_factor.

manual

A character vector naming Stan variable names which should be saved. These names should match the variable names inside the Stan code before renaming. This feature is meant for power users only and will rarely be useful outside of very special cases.

Value

A list of class "save_pars".